Create tag '3d-v3.5.0'

# Conflicts:
#	cocos/2d/category.json
#	cocos/3d/category.json
#	cocos/core/builtin/effects.ts
#	cocos/core/builtin/shader-sources/glsl4.ts
#	cocos/particle-2d/category.json
#	cocos/physics-2d/category.json
#	editor/assets/primitives.fbx.meta
#	editor/i18n/en/assets.js
#	editor/i18n/zh/assets.js
#	native/cocos/bindings/jswrapper/quickjs/SeApi.h

+++ native/cocos/bindings/manual/jsb_conversions.h
@@ -1499,7 +1499,7 @@ inline bool nativevalue_to_se(const float &from, se::Value &to, se::Object * /*c
 }
 template <>
 inline bool nativevalue_to_se(const double &from, se::Value &to, se::Object * /*ctx*/) {
-    to.setFloat(static_cast<float>(from));
+    to.setDouble(from);
     return true;
 }
 template <>
